container/airflow/requirements.txt
apache-airflow
mlflow
pandas
sqlalchemy
scikit-learn
kafka-python